% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{convertInput}
\alias{convertInput}
\title{convertInput}
\usage{
convertInput(inp, Fq_fmi = NULL, Fg_fmi = NULL, verbose = FALSE)
}
\arguments{
\item{inp}{# TODO}

\item{Fq_fmi}{# TODO}

\item{Fg_fmi}{# TODO}

\item{verbose}{TRUE will print information about the function process.
Implemented for debug purpose only.}
}
\description{
Fonction de ventilation du jeu de donnee
}
