% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{dimcst}
\alias{dimcst}
\alias{set_dimcst<-}
\title{DimCst}
\usage{
dimcst(x)

set_dimcst(x) <- value
}
\arguments{
\item{x}{object to get DimCst attributes from}
}
\description{
Get DimCst attributes from an object.
}
\details{
DimCst attributes is an integer vector of length 4
}
\author{
Maxime Jaunatre
}
