% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{fleetInput-class}
\alias{fleetInput-class}
\title{fleetInput class}
\description{
Internal class for initialising class format into list. Used in function
reformat.
}
\details{
Used by \code{reformat} function.
}
\section{Slots}{

\describe{
\item{\code{stock}}{Name of the fleet. chr.}

\item{\code{input}}{List of variables required for a fleetinside the C++ model.}
}}

