% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{format_var}
\alias{format_var}
\alias{format_vareco}
\alias{format_varsp}
\title{Format IAM.output variable as a tibble.}
\usage{
format_var(name, object)

format_vareco(name, object)

format_varsp(name, object)
}
\arguments{
\item{name}{name of the exported variable from IAM. chr.}

\item{object}{\code{\link[IAM]{iamOutput}} object created with
\code{IAM.model()}.}
}
\description{
Extract an element from \code{\link[IAM]{iamOutput}} and format this as a
tibble. This tibble is 6 column wide with specified format (see details)
}
\details{
The reconcilSPP variable return character value,
so it can't be binded with other variable which are numeric.

Format : long format for ggplot and dplyr analysis to facilitate filter and
summary computations. If the variable is not defined for a specific dimension
(column), this column is filled with NA.
\describe{
  \item{variable}{Single value repeated, but is needed if multiple variables
  are assembled with \code{bind_row()}. chr vector}
  \item{species}{Species names, can contain dynamic and static species
  depending on the variable selected.. chr vector}
  \item{fleet}{Fleet names. fct vector}
  \item{metier}{Metier names. fct vector}
  \item{age}{Ages for dynamic species. fct vector}
  \item{year}{year step for the simulation. dbl vector}
  \item{value}{Value of the variable for every column place indicated on the
  same row. dbl vector}
}
}
\author{
Maxime Jaunatre
}
