% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_args_integration.r
\name{GUIiam}
\alias{GUIiam}
\title{Graphic User Interface}
\usage{
GUIiam(INPUT, classInp = "args", desc = as.character(NA))
}
\arguments{
\item{INPUT}{# TODO named object in method}

\item{classInp}{chr between "param" and "args"}

\item{desc}{object description. chr.}
}
\description{
Graphic User Interface
}
