% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{IAM.args_scenar}
\alias{IAM.args_scenar}
\alias{IAM.args_scenar,iamArgs-method}
\title{Get Scenario arguments}
\usage{
IAM.args_scenar(object)

\S4method{IAM.args_scenar}{iamArgs}(object)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamArgs-class}} object}
}
\description{
Return the scenario arguments.
}
\examples{
data(IAM_argum_2009)
IAM.args_scenar(IAM_argum_2009)


}
\author{
Maxime Jaunatre
}
