% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{IAM.dev}
\alias{IAM.dev}
\title{Initiate a project with files and directory}
\usage{
IAM.dev(project = "Workgroup", file = "analysis.R")
}
\arguments{
\item{project}{Title of the project in which IAM is used}

\item{file}{Name of the analysis file initialised with the project.
Used as the main script of the project.
NULL will cancel the file creation.}
}
\description{
This function can be used to create a good practice environment for further
usage of the IAM package.
}
\details{
R directory is intended to store R scripts with functions.
Use \code{source("R/script.R")} to load functions inside a script.

output directory is intended to store all output created from this project.
All output can can be deleted at any point and obtained from the main script.

data contains two directory for raw and processed dataset.
Raw data should not be modified during the code evaluation.
Processed data can be deleted at any point and obtained from the main script.
If a git repository is created before calling this function,
data/raw/ directory will by default be added to .gitignore file.
}
\author{
Maxime Jaunatre
}
