% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{IAM.editArgs_Eco}
\alias{IAM.editArgs_Eco}
\alias{IAM.editArgs_Eco,iamArgs-method}
\title{Modify Economic arguments}
\usage{
IAM.editArgs_Eco(object, ...)

\S4method{IAM.editArgs_Eco}{iamArgs}(object, dr = NULL, perscCalc = NULL)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamArgs-class}} object}

\item{...}{dr and persCalc arguments}

\item{dr}{actualisation rate : numeric value in decimal
(percentage. For 4\%, dr must be 0.04)}

\item{perscCalc}{personnel cost : computation mode
\describe{
\item{0}{- constant variable}
\item{1}{- linear relation with rtbs; cshr available}
\item{2}{- linear relation with rtbs; cshr non available}
}}
}
\description{
Modify Economic arguments
}
\details{
The GUI provide many other arguments for economic parameters but they are
deprecated and it is useless to edit them
}
\examples{
data(IAM_argum_2009)
IAM_argum_2009@arguments$Eco$dr
IAM_argum_2009@arguments$Eco$perscCalc
IAM_argum_2009 <- IAM.editArgs_Eco(IAM_argum_2009, dr = 0.06, perscCalc = 0)
IAM_argum_2009@arguments$Eco$dr
IAM_argum_2009@arguments$Eco$perscCalc

}
\author{
Maxime Jaunatre
}
