% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{IAM.editArgs_Gest}
\alias{IAM.editArgs_Gest}
\alias{IAM.editArgs_Gest,iamArgs-method}
\title{Modify Gestion arguments}
\usage{
IAM.editArgs_Gest(object, ...)

\S4method{IAM.editArgs_Gest}{iamArgs}(
  object,
  active = NULL,
  control = c("Nb vessels", "Nb trips"),
  target = c("TAC", "Fbar", "TAC->Fbar"),
  espece = NULL,
  delay = NULL,
  type = c("+", "x"),
  update = NULL,
  bounds = NULL,
  tac = NULL,
  fbar = NULL,
  effSup = NULL,
  mfm = NULL
)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamArgs-class}} object}

\item{...}{selected argument}

\item{active}{Activation of the module. TRUE or FALSE input,
but accept 1 or 0 values (any numeric will be interpreted as logical)}

\item{control}{option to choose if the management will control effort
reduction by reducing the number of vessels or number of trips per vessels.}

\item{target}{management aim to reduce effort to reach TAC or Fbar,
or TAC then Fbar.}

\item{espece}{species to target for management.}

\item{delay}{option to delay the management for n years. If the delay is
larger than the year in model, the management will have no effect.}

\item{type}{option to choose if effort reduction is additive ("+") or
multiplicative ("x").}

\item{update}{option to choose is effort reduction is from t0 or t-1
iteration}

\item{bounds}{bounds for target reach at each model iteration.
num of length 2.}

\item{tac}{TAC values for each model iteration. Can be NA.}

\item{fbar}{Fbar values for each model iteration. Can be NA.}

\item{effSup}{Maximum effort per fleet and year. To limit nbds.}

\item{mfm}{Ponderation fleet/metier for effort reduction.}
}
\description{
Modify Gestion arguments
}
\examples{
data(IAM_argum_1984)
IAM.args_scenar(IAM_argum_1984)

# Activate scenario "Sc_saisine"
IAM_argum_1984 <- IAM.editArgs_Gest(IAM_argum_1984, active= TRUE)
summary(IAM_argum_1984)


}
\author{
Maxime Jaunatre
}
