% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{IAM.editArgs_Rep}
\alias{IAM.editArgs_Rep}
\alias{IAM.editArgs_Rep,iamArgs-method}
\title{Modify Replicates arguments}
\usage{
IAM.editArgs_Rep(object, ...)

\S4method{IAM.editArgs_Rep}{iamArgs}(object, n = NULL)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamArgs-class}} object}

\item{...}{selected argument}

\item{n}{Number of replicates
NULL will desactivate the Replicates module}
}
\description{
Modify Replicates arguments
}
\examples{
data(IAM_argum_1984)

# Replicate 100 times
IAM_argum_1984 <- IAM.editArgs_Rep(IAM_argum_1984, n = 100)
IAM.args_scenar(IAM_argum_1984)

# Desactivate Replicates
IAM_argum_1984 <- IAM.editArgs_Rep(IAM_argum_1984, n = NULL)
IAM.args_scenar(IAM_argum_1984)

}
\author{
Maxime Jaunatre
}
