% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_args_input_txtExport.r
\name{IAM.export}
\alias{IAM.export}
\alias{IAM.export,iamInput-method}
\alias{IAM.export,iamArgs-method}
\title{Exporting IAM objects as readable .txt file}
\usage{
IAM.export(object, folder, ...)

\S4method{IAM.export}{iamInput}(object, folder, ...)

\S4method{IAM.export}{iamArgs}(object, folder, ...)
}
\arguments{
\item{object}{TODO}

\item{folder}{The folder in which the output .txt files will be exported. Chr.}

\item{...}{deprecated argument ?}
}
\description{
Exporting IAM objects as readable .txt file
}
\details{
This method will export the input object's slots as .txt file with a generic
name corresponding to the name of the slot
("specific.txt", "input.txt","scenario.txt" & "stochastic.txt" for an
\emph{iamInput} object, or "specific.txt" & "args.txt" for an \emph{iamArgs} object.
}
\section{WARNING }{

if a file with the same name is already in the folder, it will be replaced with the exported file. So, use it with caution!!
}

