% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.r
\name{IAM.format_quant}
\alias{IAM.format_quant}
\title{Simplify multiple simulations}
\usage{
IAM.format_quant(var_format, probs = c(0.25, 0.75), select_indiv = 1)
}
\arguments{
\item{var_format}{Tibble or data.frame format produced by
\code{\link[IAM]{IAM.format}}.}

\item{probs}{Quantile arguments for the values distribution once grouped.
Vector of 2 numeric values in [0, 1].}

\item{select_indiv}{Single numeric value which allow to conserve a single
simulation value for example. Default is 1.}
}
\value{
Format : long format for ggplot and dplyr analysis to facilitate filter and
summary computations. If the variable is not defined for a specific dimension
(column), this column is filled with NA.
\describe{
  \item{sim_name}{Simulation name. chr.}
  \item{variable}{Single value repeated, but is needed if multiple variables
  are assembled with \code{bind_row()}. chr vector}
  \item{species}{Species names, can contain dynamic and static species
  depending on the variable selected.. chr vector}
  \item{fleet}{Fleet names. fct vector}
  \item{metier}{Metier names. fct vector}
  \item{age}{Ages for dynamic species. fct vector}
  \item{year}{year step for the simulation. dbl vector}
  \item{quant1}{First quantile from \code{probs} argument.}
  \item{quant2}{Second quantile from \code{probs} argument.}
  \item{median}{Median computed from the grouped values.}
  \item{value}{Value of the variable for every column place indicated on the
  same row. dbl vector}
}
}
\description{
Simplify multiple simulations and group them after computing
quantile and median value (per species, fleet, metier, age, year).
A selected simulation is kept to illustrate a possible way.
}
\examples{
library(dplyr)
library(magrittr)

data("IAM_input_1984")
data("IAM_argum_1984")
sim_statu_quo <- IAM::IAM.model(objArgs = IAM_argum_1984, objInput = IAM_input_1984)
res1 <- IAM.format(sim_statu_quo, c("SSB"), n = 1) \%>\%
  dplyr::filter(species == "ARC", year < 1987)
res2 <- mutate(res1, n = 2, value = value + rnorm(1, sd = 100))
res <- rbind(res1, res2)

IAM.format_quant(res, c(0.025, 0.975), 2)

}
\author{
Maxime Jaunatre
}
