% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{IAM.input}
\alias{IAM.input}
\alias{IAM.input,character,missing,missing,missing-method}
\alias{IAM.input,character,character,missing,missing-method}
\alias{IAM.input,character,character,character,missing-method}
\alias{IAM.input,character,character,missing,character-method}
\alias{IAM.input,character,character,character,character-method}
\title{'iamInput' objects creator}
\usage{
IAM.input(fileIN, fileSPEC, fileSCEN, fileSTOCH, ...)

\S4method{IAM.input}{character,missing,missing,missing}(
  fileIN,
  t_init,
  nbStep = 20,
  t_hist_max = t_init,
  desc = "My Input",
  folderFleet = NULL,
  readXLS = "openxlsx",
  Fq_i = NULL,
  iniFq_i = NULL,
  Fq_fmi = NULL,
  iniFq_fmi = NULL,
  FqLwt_i = NULL,
  iniFqLwt_i = NULL,
  FqLwt_fmi = NULL,
  iniFqLwt_fmi = NULL,
  FqDwt_i = NULL,
  iniFqDwt_i = NULL,
  FqDwt_fmi = NULL,
  iniFqDwt_fmi = NULL,
  Nt0s1q = NULL,
  Ni0q = NULL,
  iniNt0q = NULL,
  matwt = NULL,
  Fg_fmi = NULL,
  dg_fmi = NULL,
  ...
)

\S4method{IAM.input}{character,character,missing,missing}(fileIN, fileSPEC, desc = "My Input", ...)

\S4method{IAM.input}{character,character,character,missing}(fileIN, fileSPEC, fileSCEN, desc = "My Input", ...)

\S4method{IAM.input}{character,character,missing,character}(fileIN, fileSPEC, fileSTOCH, desc = "My Input", ...)

\S4method{IAM.input}{character,character,character,character}(fileIN, fileSPEC, fileSCEN, fileSTOCH, desc = "My Input", ...)
}
\arguments{
\item{fileIN}{.xls standard IAM input file}

\item{fileSPEC}{deprecated argument.}

\item{fileSCEN}{deprecated argument.}

\item{fileSTOCH}{deprecated argument.}

\item{...}{Further arguments used when the input is a .xls. Described below.}

\item{t_init}{Only for first signature. Initial year.}

\item{nbStep}{Only for first signature. Number of timesteps (including initial year).}

\item{t_hist_max}{Only for first signature. Last year considered for 'historical' slot.}

\item{desc}{Object descriptor (default value : "My Input").}

\item{folderFleet}{Folder containing fleets input sheets (Optionnal. Default value : NULL).}

\item{readXLS}{Character. Package to be used to read parameter Excel file. # TODO remove
'openxlsx' for post-2007 excel version (.xlsx files),
'XLConnect' for pre-2007 excel version (.xls files).(Default value : "openxlsx").}

\item{Fq_i}{List containing SS3 parameters arrays for quarterly stock dynamic simulation.
One element per considered species, all names must match with input species.
Fishing mortality per season, morph, and age.
(Optionnal. Default value : NULL. Example : list(hake=array(...,dim=c(4,4,nAge))).}

\item{Fq_fmi}{See above. Fishing mortality per season, morph, fleet, "economic" metiers and age
(Optionnal. Default value : NULL).}

\item{Ni0q}{See above. Recruits numbers per season (Optionnal. Default value : NULL).}
}
\description{
'iamInput' objects creator
}
