% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_args_integration.r
\name{IAM.input2args}
\alias{IAM.input2args}
\alias{IAM.input2args,iamInput-method}
\title{'iamArgs' objects creator by code}
\usage{
IAM.input2args(object, ...)

\S4method{IAM.input2args}{iamInput}(object, desc = NULL)
}
\arguments{
\item{object}{An \code{\link[IAM]{iamInput-class}} or \code{\link[IAM]{iamArgs-class}} object.}

\item{...}{desc argument}

\item{desc}{Object descriptor (default value : \code{as.character(NA)}).
If not provided, copied the description slot of object.}
}
\description{
'iamArgs' objects creator by code
}
