% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_simulation.r
\docType{methods}
\name{IAM.model-methods}
\alias{IAM.model-methods}
\alias{IAM.model}
\alias{IAM.model,iamArgs,iamInput-method}
\title{Modelization method}
\usage{
IAM.model(objArgs, objInput, ...)

\S4method{IAM.model}{iamArgs,iamInput}(
  objArgs,
  objInput,
  desc = as.character(NA),
  mOTH = 0,
  updateE = 0,
  TACbyF = NULL,
  TACtot = NULL,
  Ftarg = NULL,
  W_Ftarg = NULL,
  MeanRec_Ftarg = NULL,
  TACbyFoptimCTRL = list(maxIter = as.integer(7), diffZmax = 1e-04, lambda = 0.9,
    t_stop = 0),
  recList = list(),
  recParamList = list(),
  ParamSPMList = list(),
  parBehav = list(active = as.integer(0), type = as.integer(3), FMT = NULL, MU = NULL,
    MUpos = as.integer(0), ALPHA = NULL),
  parOptQuot = list(active = as.integer(0), pxQuIni = NULL, pxQuMin = NULL, pxQuMax =
    NULL, lambda = 0.1, sdmax = 0, ftol = 1e-07, itmax = 500),
  tacControl = list(tolVarTACinf = NA, tolVarTACsup = NA, corVarTACval = NA,
    corVarTACnby = 2, Blim = NA, Bmax = NA, BlimTrigger = as.integer(0), typeMng = NA),
  stochPrice = list(),
  parOQD = list(activeQR = as.integer(0), listQR = NULL, listQR_f = NULL),
  verbose = FALSE,
  force_t = NULL,
  ...
)
}
\arguments{
\item{objArgs}{output from \code{\link[IAM]{IAM.args-methods}}}

\item{objInput}{output from \code{\link[IAM]{IAM.input-methods}}}

\item{...}{there is so much more arg.... # TODO}

\item{verbose}{Show messages to follow the process. Mainly used for debug process.}
}
\description{
Modelization method
}
