% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.r
\name{IAM.bwplot}
\alias{IAM.bwplot}
\alias{IAM.zone}
\alias{IAM.barIC}
\alias{IAM.barplot}
\alias{IAM.plot}
\title{Plotting functions for IAM package based on lattice methods}
\usage{
IAM.bwplot(formula, data, ...)

IAM.zone(formula, data, ...)

IAM.barIC(formula, data, ...)

IAM.barplot(formula, data, ...)

IAM.plot(formula, data, ...)
}
\arguments{
\item{formula}{Typical 'Lattice' formula. No 'groups' parameter,
so multiple terms separated by a 'plus' sign is required for grouping
(cf \emph{Extended formula interface} part in 'xyplot' documentation)}

\item{data}{A data frame containing values for any variables in the formula
(for example, an \emph{IAM.unite} output)}

\item{...}{Further graphical parameters (see below for details)}
}
\description{
TODO
}
\details{
Graphical parameters used are :
 \describe{
\item{type}{Character}
\item{pch}{Numeric}
\item{cex.pch}{Numeric}
\item{lty}{Numeric}
\item{lwd}{Numeric}
\item{relation}{Character}
\item{key}{Logical}
\item{pch.leg}{Numeric}
\item{pchSize.leg}{Numeric}
\item{lty.leg}{Numeric}
\item{lwd.leg}{Numeric}
\item{txt.leg}{Character}
\item{font.leg}{Numeric}
\item{cex.leg}{Numeric}
\item{space}{Character}
\item{alpha}{Numeric}
\item{add.v}{Numeric}
\item{add.h}{Numeric}
\item{add.lwd}{Numeric}
\item{add.lty}{Numeric}
\item{add.col}{Character}
\item{cex.lab.x}{Numeric}
\item{cex.lab.y}{Numeric}
\item{font.lab.x}{Numeric}
\item{font.lab.y}{Numeric}
\item{cex.axis.x}{Numeric}
\item{cex.axis.y}{Numeric}
\item{font.axis.x}{Numeric}
\item{font.axis.y}{Numeric}
\item{rot.x}{Numeric}
\item{rot.y}{Numeric}
\item{cex.strip}{Numeric}
\item{font.strip}{Numeric}
\item{col.strip}{Character}
\item{col}{Character}
\item{fill}{Character}
\item{xlab}{Character}
\item{ylab}{Character}
\item{origin}{Numeric}
\item{as.table}{Logical}
 }
}
