% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_0_utils_test_plot.R
\name{IAM.test_plot}
\alias{IAM.test_plot}
\title{Module shiny pour l'explo graphique.}
\usage{
IAM.test_plot(object)
}
\arguments{
\item{object}{Table from IAM simulation. Of class "iam_quantbl" or
"iam_formtbl" (second one will be modified with IAM.format_quant function).}
}
\description{
Permet de faire les plots de Fbar, SSB et L_et en fonction du temps.
}
\author{
Maxime Jaunatre
}
