% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.r
\name{IAM.unite}
\alias{IAM.unite}
\alias{IAM.unite,list-method}
\title{Formating method for grouping various \code{\link[IAM]{iamOutput-class}} or \code{\link[IAM]{iamOutputRep-class}} objects}
\usage{
IAM.unite(listObj, ...)

\S4method{IAM.unite}{list}(listObj, ...)
}
\arguments{
\item{listObj}{A list containing \code{\link[IAM]{iamOutput-class}} OR \code{\link[IAM]{iamOutputRep-class}} objects. Mixed lists are not allowed.}

\item{...}{Further arguments :
\describe{
\item{elmt}{ Name of the operating variable. chr.}
\item{spp}{Name of the considered species (relevant only for some variables). chr}
\item{agg}{Index(es) describing dimensions on which agregation must be done,
referring to 'DimCst' attributes (1=fleet, 2=metier, 3=age/category, 4=time). num.}
\item{headers}{Optionnal. Dimension that will be developped as headers in the output data.frame. chr.}
\item{subs}{An optionnal 'subset' argument to operate a final subsetting on output dataframe. logic}
}}
}
\description{
Formating method for grouping various \code{\link[IAM]{iamOutput-class}} or \code{\link[IAM]{iamOutputRep-class}} objects
}
