% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_service_specific_credential}
\alias{iam_delete_service_specific_credential}
\title{Deletes the specified service-specific credential}
\usage{
iam_delete_service_specific_credential(UserName,
  ServiceSpecificCredentialId)
}
\arguments{
\item{UserName}{The name of the IAM user associated with the service-specific
credential. If this value is not specified, then the operation assumes
the user whose credentials are used to call the operation.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: \_+=,.@-}

\item{ServiceSpecificCredentialId}{[required] The unique identifier of the service-specific credential. You can get
this value by calling ListServiceSpecificCredentials.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters that
can consist of any upper or lowercased letter or digit.}
}
\description{
Deletes the specified service-specific credential.
}
\section{Request syntax}{
\preformatted{iam$delete_service_specific_credential(
  UserName = "string",
  ServiceSpecificCredentialId = "string"
)
}
}

\keyword{internal}
