% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{IAM_example}
\alias{IAM_example}
\title{Path to the IAM extdata files}
\usage{
IAM_example(file)
}
\arguments{
\item{file}{Name of the file.
Need to be in inst/extdata/ directory of the package}
}
\description{
Allorw to get paths from the package installation location for
all extdata files.
}
\details{
simplified system from {readr} package.
}
\examples{
IAM_example("IAM_SS3_2009.RData")
IAM_example("inputFile.xlsx")
IAM_example("fleets")

}
