% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_generate_credential_report}
\alias{iam_generate_credential_report}
\title{Generates a credential report for the AWS account}
\usage{
iam_generate_credential_report()
}
\description{
Generates a credential report for the AWS account. For more information
about the credential report, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/credential-reports.html}{Getting Credential Reports}
in the \emph{IAM User Guide}.
}
\keyword{internal}
