% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.r
\name{IAM_theme}
\alias{IAM_theme}
\title{Theme for IAM plots}
\usage{
IAM_theme(blue = TRUE)
}
\arguments{
\item{blue}{modify the background for a blue color. TRUE by default.
False will desactivate for a white background.}
}
\description{
Modify background for IAM plots.
}
\author{
maxime jaunate
}
