% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.r
\name{iamFormat}
\alias{iamFormat}
\title{Formating method}
\usage{
iamFormat(
  object,
  elmt = character(),
  spp = character(),
  agg = NA,
  headers = as.character(NA),
  subs,
  index = NA
)
}
\arguments{
\item{object}{An \code{\link[IAM]{iamOutput-class}} or \code{\link[IAM]{iamOutputRep-class}} object.}

\item{elmt}{Name of the operating variable. chr (1).}

\item{spp}{Name of the considered species (relevant only for some variables). chr (1).}

\item{agg}{Index(es) describing dimensions on which agregation must be done,
referring to 'DimCst' attributes (1=fleet, 2=metier, 3=age/category, 4=time). num (<4)}

\item{headers}{Optionnal. Dimension that will be developped as headers in
the output data.frame. chr.}

\item{subs}{An optionnal 'subset' argument to operate a final subsetting on output dataframe. chr (1)}

\item{index}{# TODO argum is ?}
}
\value{
data.frame avec colonnes : value + dimensions
}
\description{
fonction interne d'agregation et de formatage de la donnee a representer
selon divers parametres
}
