% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_output_objects.r
\docType{class}
\name{iamOutput-class}
\alias{iamOutput-class}
\title{Class "iamOutput"}
\description{
Class "iamOutput"
}
\section{Slots}{

\describe{
\item{\code{desc}}{Short description of the object. chr.}

\item{\code{arguments}}{# TODO}

\item{\code{specific}}{# TODO}

\item{\code{outputSp}}{# TODO}

\item{\code{output}}{# TODO}
}}

\examples{
showClass("iamOutput")


}
\author{
Mathieu Merzereaud
}
