% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_output_objects.r
\docType{class}
\name{iamOutputRep-class}
\alias{iamOutputRep-class}
\title{Class "iamOutputRep"}
\description{
Class "iamOutputRep"
}
\section{Slots}{

\describe{
\item{\code{desc}}{Copy of the desc slot from \code{\link[IAM]{iamInput-class}}}

\item{\code{arguments}}{Copy of the arguments slot from \code{\link[IAM]{iamArgs-class}}}

\item{\code{specific}}{Copy of the specific slot from \code{\link[IAM]{iamInput-class}}}

\item{\code{outputSp}}{List of biologic variable. Each element is a list for a
variable (example F) formated as an element per dynamic stock.}

\item{\code{output}}{List of economic variable. Each element is a table with various
format depending of dimension (fleet, metier, year)}
}}

\examples{
showClass("iamOutputRep")


}
\author{
Mathieu Merzereaud
}
