\name{iamToFLR}
\alias{iamToFLR}
\alias{iamToFLR-methods}
\alias{iamToFLR,iamOutput,iamInput-method}
\alias{iamToFLR,iamOutputRep,iamInput-method}
\docType{methods}

\title{Conversion process from IAM objects to FLR objects}
\description{ToDo}

\usage{
iamToFLR(object,input,type="Stock",\dots)
}

\arguments{
  \item{object}{ToDo.}
  \item{input}{ToDo.}
  \item{type}{Character. Type of output FLR object. To be chosen between "Stock", "Catch", "Fleet" & "FleetMetier".}
  \item{\dots}{Further arguments}  
}


\section{Methods}{
  \describe{
	\item{iamToFLR}{\code{signature(object = "iamOutput", input = "iamInput")}}
	\item{iamToFLR}{\code{signature(object = "iamOutputRep", input = "iamInput")}}
}}


\keyword{methods}


