% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{init_listHisto}
\alias{init_listHisto}
\title{init_listHisto}
\usage{
init_listHisto(List, t_init, t_hist_max)
}
\arguments{
\item{List}{a result table filtered before}

\item{t_hist_max}{first time of the modelisation}
}
\description{
initialise list Historique
}
