% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{marketInput-class}
\alias{marketInput-class}
\title{marketInput Class}
\description{
Internal class for initialising class format into list. Used in function
reformat.
}
\details{
Used by \code{reformat} function.
}
\section{Slots}{

\describe{
\item{\code{input}}{List of variables required for marketinside the C++ model.}
}}

