% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{read.Pflex}
\alias{read.Pflex}
\title{read.Pflex}
\usage{
read.Pflex(file, nam_stock, nam_stock_bis)
}
\arguments{
\item{file}{excell file (.xlsx) imported by IAM.input.
Must contain a Price_flexibility sheet.Character string.}

\item{nam_stock}{vector of character string with stock names.
Format is "Stock__X", X being an abreviation for a precise stock.}

\item{nam_stock_bis}{# TODO what is this format

# TODO add example file.}
}
\description{
Function to import a Price_flexibility sheet from the input file.
}
\author{
Florence Briton 05/2019
}
