% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{read.Scenar}
\alias{read.Scenar}
\title{read.Scenarii}
\usage{
read.Scenar(file)
}
\arguments{
\item{file}{excell file (.xlsx) imported by IAM.input.
Must contain a Scenarii sheet.Character string.}
}
\description{
read.Scenarii
}
