% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{read.sheet}
\alias{read.sheet}
\title{read.sheet}
\usage{
read.sheet(file, sheet)
}
\arguments{
\item{file}{excell file (.xlsx) imported by IAM.input.
Must contain a sheet named after the second parameter.Character string.}

\item{sheet}{sheet name inside the file imported by IAM.input.
Character string.}
}
\description{
Import a sheet from an excell file.
}
\details{
Replace "," with "." and all columns into character.
Replace NA values with empty string ("")
}
