% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{result_filtre}
\alias{result_filtre}
\title{Reapeated code that I don't understand yet}
\usage{
result_filtre(result, indEmpt)
}
\arguments{
\item{result}{a readed sheet from a file, modified before}

\item{indEmpt}{something I need to understand # TODO}
}
\description{
Idea is DRY (Don't Repeat Yourself)
}
