% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_input_integration.r
\name{standFormat}
\alias{standFormat}
\title{standFormat}
\usage{
standFormat(DF, nbStep, modF, modM, modI, modC, alk, as.na = NULL)
}
\description{
reformat a dataframe into a standart formated matrix
}
