% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{staticStockInput-class}
\alias{staticStockInput-class}
\title{staticStockInput Class}
\description{
Internal class for initialising class format into list. Used in function
reformat.
}
\details{
Used by \code{reformat} function.
}
\section{Slots}{

\describe{
\item{\code{stock}}{Name of the stock. chr.}

\item{\code{input}}{List of variables required for a stock inside the C++ model.}
}}

