% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{summary,iamArgs-method}
\alias{summary,iamArgs-method}
\title{summary method for iamInput class}
\usage{
\S4method{summary}{iamArgs}(object, ...)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamArgs-class}} object}

\item{...}{Unused argument}
}
\description{
summary method for iamInput class
}
\details{
The GUI provide many other arguments for economic parameters but they are
deprecated and it is useless to show them
}
\examples{
data("IAM_argum_2009")
IAM_argum_2009@arguments$Gestion$active <- 1
IAM_argum_2009@arguments$Replicates$active <- 1
IAM_argum_2009@arguments$Scenario$ALLscenario <- c("Scenario1", "Scenario2")
IAM_argum_2009 <- IAM.editArgs_Scenar(IAM_argum_2009, selected = 1)
summary(IAM_argum_2009)

}
\author{
Maxime Jaunatre
}
