% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_Edit_iamArgs.R
\name{summary,iamInput-method}
\alias{summary,iamInput-method}
\title{summary method for iamInput class}
\usage{
\S4method{summary}{iamInput}(object, ...)
}
\arguments{
\item{object}{a \code{\link[IAM]{iamInput-class}} object}

\item{...}{Unused argument}
}
\description{
summary method for iamInput class
}
\examples{
data("IAM_input_1984")
summary(IAM_input_1984)

}
\author{
Maxime Jaunatre
}
