% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\name{val.iamArgs}
\alias{val.iamArgs}
\title{Checking validity of iamArgs Class}
\usage{
val.iamArgs(object)
}
\arguments{
\item{object}{\code{\link[IAM]{iamArgs-class}} object}
}
\description{
Checking validity of iamArgs Class
}
