% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_args_integration.R, R/2_args_integration.r
\name{IAM.input2args}
\alias{IAM.input2args}
\alias{IAM.input2args,iamInput-method}
\alias{IAM.args}
\alias{IAM.args,iamInput,missing-method}
\alias{IAM.args,character,character-method}
\alias{IAM.args,iamArgs,missing-method}
\title{'iamArgs' objects creator}
\usage{
IAM.input2args(object, desc = NULL, ...)

\S4method{IAM.input2args}{iamInput}(object, desc = NULL, ...)

IAM.args(object, specific, ...)

\S4method{IAM.args}{iamInput,missing}(object, desc = as.character(NA), ...)

\S4method{IAM.args}{character,character}(object, specific, desc = as.character(NA), ...)

\S4method{IAM.args}{iamArgs,missing}(object, desc = as.character(NA), ...)
}
\arguments{
\item{object}{TODO}

\item{desc}{Object descriptor (default value : \code{as.character(NA)}).
If not provided, copied the description slot of object.}

\item{...}{# TODO : this is for the specific file.}

\item{specific}{TODO}
}
\description{
'iamArgs' objects creator

'iamArgs' objects creator

Etape d'initialisation

Etape d'initialisation from txt file.

Etape de modification
To update an iamArgs object, the function require to have original
information about input. This is why at the end of initialization
the specific slot is copied from iamInput to iamArgs
}
