% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_args_integration.r
\name{IAM.args}
\alias{IAM.args}
\alias{IAM.args,iamInput,missing-method}
\alias{IAM.args,character,character-method}
\alias{IAM.args,iamArgs,missing-method}
\title{'iamArgs' objects creator}
\usage{
IAM.args(object, specific, ...)

\S4method{IAM.args}{iamInput,missing}(object, desc = as.character(NA), ...)

\S4method{IAM.args}{character,character}(object, specific, desc = as.character(NA), ...)

\S4method{IAM.args}{iamArgs,missing}(object, desc = as.character(NA), ...)
}
\arguments{
\item{object}{TODO}

\item{specific}{TODO}

\item{...}{TODO}

\item{desc}{Object descriptor (default value : \code{as.character(NA)})}
}
\description{
'iamArgs' objects creator
}
