% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.r
\name{IAM.format-method}
\alias{IAM.format-method}
\alias{IAM.format}
\alias{IAM.format,iamOutput,character-method}
\alias{IAM.format,iamOutput-method}
\alias{format_vareco}
\alias{format_varsp}
\title{Format IAM.output variable as a tibble.}
\usage{
IAM.format(object, name, sim_name = NA_character_, n = NA_real_)

\S4method{IAM.format}{iamOutput,character}(object, name, sim_name = NA_character_, n = NA_real_)

format_vareco(object, name, sim_name = NA_character_, n = NA_real_)

format_varsp(object, name, sim_name = NA_character_, n = NA_real_)
}
\arguments{
\item{object}{\code{\link[IAM]{iamOutput-class}} object created with
\code{IAM.model()}.}

\item{name}{names of the exported variable from IAM. "summary" will select
few variables that are usually displayed for most common scenarii ("Fbar",
"SSB", "L_et", "N", "nbv_f", "effort2_f", "GVLav_f",
"gva_f", "gp_f", "wageg_f", "wagen_f"). chr.}

\item{sim_name}{Name of the simulation, used for joining multiple scenarii
iamOutput. NA by default, chr.}

\item{n}{Number of the simulation, used for joining replicated iamOutput.
NA by default, dbl.}
}
\value{
Format : long format for ggplot and dplyr analysis to facilitate filter and
summary computations. If the variable is not defined for a specific dimension
(column), this column is filled with NA.
\describe{
  \item{sim_name}{Simulation name. chr.}
  \item{n}{Number of the simulation. dbl}
  \item{variable}{Single value repeated, but is needed if multiple variables
  are assembled with \code{bind_row()}. chr vector}
  \item{species}{Species names, can contain dynamic and static species
  depending on the variable selected.. chr vector}
  \item{fleet}{Fleet names. fct vector}
  \item{metier}{Metier names. fct vector}
  \item{age}{Ages for dynamic species. fct vector}
  \item{year}{year step for the simulation. dbl vector}
  \item{value}{Value of the variable for every column place indicated on the
  same row. dbl vector}
}
}
\description{
Extract an element from \code{\link[IAM]{iamOutput-class}} and format this as a
tibble. This tibble is 6 column wide with specified format (see details)
}
\details{
The reconcilSPP variable return character value,
so it can't be binded with other variable which are numeric.
}
\examples{
data("IAM_input_2009")
data("IAM_argum_2009")
sim_statu_quo <- IAM::IAM.model(objArgs = IAM_argum_2009, objInput = IAM_input_2009)

#' cas vide return NULL
IAM.format(sim_statu_quo, "not a variable")
#' cas simple bio
IAM.format(sim_statu_quo, "SSB")
#' cas complexe bio
IAM.format(sim_statu_quo, "F")
IAM.format(sim_statu_quo, "N_S1M1")
#' cat eco
IAM.format(sim_statu_quo, name = "ratio_gp_K_f")
#' Warning here values are string !
IAM.format(sim_statu_quo, name = "reconcilSPP")



}
\author{
Maxime Jaunatre
}
