% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\docType{package}
\name{IAM}
\alias{IAM}
\title{Impact Assesment Model}
\description{
IAM (Impact Assessment Model) is a bio-economic model
for the simulation of fisheries dynamics,
integrating specific decision support tools within the
framework of the theoretical implementation of management measures
It is a discrete time, multi-fleet, multi-trade, multi-species annual
bio-economic model with "age" components for the biological part,
and "commercial category" components for the economic part.
}
