% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{iamArgs-class}
\alias{iamArgs-class}
\title{iamArgs class object}
\description{
# TODO
}
\details{
Used by \code{IAM.Args} method that use tcltk package for GUI.
}
\section{Slots}{

\describe{
\item{\code{desc}}{# TODO descr}

\item{\code{arguments}}{# TODO descr}

\item{\code{specific}}{# TODO descr}
}}

