% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{iamArgs-class}
\alias{iamArgs-class}
\title{Class "iamArgs"}
\description{
# TODO
}
\details{
Used by \code{IAM.Args} method that use tcltk package for GUI.
}
\section{Slots}{

\describe{
\item{\code{desc}}{Copy of the desc argument from \code{\link[IAM]{iamInput-class}}}

\item{\code{arguments}}{# TODO descr}

\item{\code{specific}}{Copy of the specific argument from \code{\link[IAM]{iamInput-class}}}
}}

\examples{
showClass("iamArgs")

}
\author{
Mathieu Merzereaud
}
