% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{iamInput-class}
\alias{iamInput-class}
\title{iamInput Class}
\description{
# TODO
}
\details{
The \code{iamInput} class has methods defined for creation of
a new object, usage in \code{iamArgs} initiation and in \code{IAM.Model}.
}
\section{Slots}{

\describe{
\item{\code{desc}}{# TODO descr}

\item{\code{specific}}{# TODO descr}

\item{\code{historical}}{# TODO descr}

\item{\code{input}}{# TODO descr}

\item{\code{scenario}}{# TODO descr}

\item{\code{stochastic}}{# TODO descr}

\item{\code{optimization}}{# TODO descr}
}}

