% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_output_objects.r
\docType{class}
\name{iamOutput-class}
\alias{iamOutput-class}
\title{Class "iamOutput"}
\description{
Class "iamOutput"
}
\section{Slots}{

\describe{
\item{\code{desc}}{Copy of the desc slot from \code{\link[IAM]{iamInput-class}}}

\item{\code{arguments}}{Copy of the arguments slot from \code{\link[IAM]{iamArgs-class}}}

\item{\code{specific}}{Copy of the specific slot from \code{\link[IAM]{iamInput-class}}}

\item{\code{outputSp}}{# TODO}

\item{\code{output}}{# TODO}
}}

\examples{
showClass("iamOutput")


}
\author{
Mathieu Merzereaud
}
