% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_input_objects.r
\docType{class}
\name{marketInput-class}
\alias{marketInput-class}
\title{marketInput Class}
\description{
# TODO
}
\details{
Used by \code{reformat} function. # TODO find in what extent
}
\section{Slots}{

\describe{
\item{\code{input}}{# TODO descr}
}}

